Eccentricity distribution, output, aborting bad fits
Allows a range of eccentricities to be used in a multiple system
simulation. Now can output textfiles with the input/output, and can save
the simulator object itself as a pickle for later reference. Simulation
can abort bootstrapping for uncertainties if the fit is already "bad"
(period is more than 10% off)